+++ dynamic-dialog-mixin.js
@@ -12,11 +12,12 @@ import './etools-dialog.js';
  * @mixinFunction
  * @appliesMixin EtoolsLogsMixin
  * @demo demo/index-dynamic.html
+ * !!! DEPRECATED - use the exported methods form dynamic-dialog.js
  */
 export const DynamicDialogMixin = baseClass => class extends EtoolsLogsMixin(baseClass) {
 
   createDialog(title, size, okBtnText, cancelBtnText, closeCallback, content, removePadding, theme) {
-    let config = {
+    const config = {
       title: title,
       size: size,
       okBtnText: okBtnText,
@@ -26,7 +27,7 @@ export const DynamicDialogMixin = baseClass => class extends EtoolsLogsMixin(bas
       noPadding: removePadding,
       theme: theme
     };
-    let dialog = this.createDynamicDialog(config);
+    const dialog = this.createDynamicDialog(config);
     document.querySelector('body').appendChild(dialog);
     return dialog;
   }
@@ -36,10 +37,10 @@ export const DynamicDialogMixin = baseClass => class extends EtoolsLogsMixin(bas
       return null;
     }
 
-    let dialog = document.createElement('etools-dialog');
+    const dialog = document.createElement('etools-dialog');
     this._applyDefaultDialogConfig(dialog);
 
-    for (let propertyName in config) {
+    for (const propertyName in config) {
       if (!config.hasOwnProperty(propertyName) || propertyName === 'closeCallback') {
         continue;
       }
@@ -60,7 +61,7 @@ export const DynamicDialogMixin = baseClass => class extends EtoolsLogsMixin(bas
 
     document.querySelector('body').appendChild(dialog);
 
-    let msgPlaceholder = dialog.shadowRoot.querySelector('#dynamicContent');
+    const msgPlaceholder = dialog.shadowRoot.querySelector('#dynamicContent');
     msgPlaceholder.appendChild(config.content);
     return dialog;
   }
