+++ index.js
@@ -1,0 +1,4 @@
+import EtoolsDialog from './EtoolsDialog.js';
+
+export {createDynamicDialog, createDialog} from './dynamic-dialog.js';
+export {EtoolsDialog} from './EtoolsDialog.js';
