+++ package.json
@@ -1,15 +1,15 @@
 {
   "name": "@unicef-polymer/etools-dialog",
   "description": "Simple dialog element",
-  "version": "4.0.0",
+  "version": "4.1.0-rc.1",
   "homepage": "https://github.com/unicef-polymer/etools-dialog",
   "repository": {
     "type": "git",
     "url": "https://github.com/unicef-polymer/etools-dialog.git"
   },
   "license": "Apache-2.0",
-  "main": "etools-dialog.js",
-  "module": "etools-dialog.js",
+  "main": "index.js",
+  "module": "index.js",
   "devDependencies": {
     "@polymer/gen-typescript-declarations": "^1.6.2",
     "@polymer/iron-component-page": "^4.0.0",
